Split components in folders, update api endpoints and import statements
