+++ app/manager-frontend/src/components/activitysources/ActivitySourceSummary.jsx
