+++ app/manager-frontend/src/components/resultsmodifiers/ResultsModifier.jsx
